Rename gonamekoclient to gonameko
