+++ go.mod
@@ -1,8 +1,9 @@
-module github.com/iamdavidzeng/gonamekoclient
+module github.com/iamdavidzeng/gonameko
 
-go 1.13
+go 1.16
 
 require (
 	github.com/satori/go.uuid v1.2.0
 	github.com/streadway/amqp v1.0.0
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
