+++ utils.go
@@ -1,0 +1,9 @@
+package gonameko
+
+import "log"
+
+func FailOnError(err error, msg string) {
+	if err != nil {
+		log.Fatalf("%s: %s", msg, err)
+	}
+}
